move zoom to base class
