+++ Integreat/Integreat.Droid/CustomRenderer/PdfWebViewRenderer.cs
@@ -1,5 +1,4 @@
-﻿using System.ComponentModel;
-using Integreat.Droid.CustomRenderer;
+﻿using Integreat.Droid.CustomRenderer;
 using Integreat.Shared.CustomRenderer;
 using Xamarin.Forms;
 using Xamarin.Forms.Platform.Android;
@@ -12,7 +11,7 @@ namespace Integreat.Droid.CustomRenderer
     /// </summary>
     /// <seealso cref="Xamarin.Forms.Platform.Android.WebViewRenderer" />
     /// <inheritdoc />
-    public class PdfWebViewRenderer : WebViewRenderer
+    public class PdfWebViewRenderer : ZoomingWebViewRenderer
     {
         protected override void OnElementChanged(ElementChangedEventArgs<WebView> e)
         {
@@ -36,18 +35,6 @@ namespace Integreat.Droid.CustomRenderer
                 Control.LoadUrl($"file:///android_asset/web/viewer.html?file={pdfWebView.Uri}");
                 Control.Reload();
             }
-        }
-
-        protected override void OnElementPropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            //check if source is external
-            if (Control != null && ((WebView)sender).Source is UrlWebViewSource)
-            {
-                //enable Android zoom
-                Control.Settings.BuiltInZoomControls = true;
-                Control.Settings.DisplayZoomControls = false;
-            }
-            base.OnElementPropertyChanged(sender, e);
-        }
+        } 
     }
 }
