+++ Integreat/Integreat.Droid/Integreat.Droid.csproj
@@ -259,6 +259,7 @@
     <Compile Include="CustomRenderer\GeneralWebViewRenderer.cs" />
     <Compile Include="CustomRenderer\PdfWebViewRenderer.cs" />
     <Compile Include="CustomRenderer\SpaceFittingLabelRenderer.cs" />
+    <Compile Include="CustomRenderer\ZoomingWebViewRenderer.cs" />
     <Compile Include="Effects\AndroidStatusBarEffect.cs" />
     <Compile Include="Helpers\Globals.cs" />
     <Compile Include="Helpers\Settings.cs" />
